refactor: extract screenshot support from Renderer and fix rescue modifier
Extract screenshot capture methods (maybe_save_screenshot,
convert_bgra_to_rgba, validate_screenshot_path) into
Renderer::Screenshot module. Fix rescue modifier to use
begin...rescue...end block and remove rubocop:disable comment.

Refs #426
